Get all dataset tables

Download data from the dataset ("dataset"), observation ("observacao"), layer ("camada"), and metadata ("metadado") tables of a datasets contained in the Free Brazilian Repository for Open Soil Data -- febr, http://www.ufsm.br/febr.

febr(dataset, merge = FALSE, progress = TRUE, verbose = TRUE, ...)

Arguments

dataset
Character vector indicating one or more datasets. Identification codes should be as recorded in http://www.ufsm.br/febr/catalog/. Use dataset = "all" to download all datasets.

merge
(optional) Logical value indicating if the observation ("observacao") and layer ("camada") tables should be merged. Defaults to merge = FALSE. See merge for more details.

progress
(optional) Logical value indicating if a download progress bar should be displayed.

verbose
(optional) Logical value indicating if informative messages should be displayed. Generally useful to identify datasets with inconsistent data. Please report to febr-forum@googlegroups.com if you find any issue.

...
(optional) Arguments passed to observation and layer.

Value

A list of data frames with data on the chosen dataset.

See also

Examples

res <- febr(dataset = "ctb0003", merge = TRUE, variable = "all")

#> Downloading table ctb0003-dataset...

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> Downloading table ctb0003-observacao...

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> Downloading table ctb0003-camada...

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> Downloading table ctb0003-metadado...

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

str(res)

#> List of 3
#>  $ dataset :'data.frame':	25 obs. of  2 variables:
#>   ..$ campo: chr [1:25] "dataset_id" "dataset_titulo" "dataset_descricao" "dataset_versao" ...
#>   ..$ valor: chr [1:25] "ctb0003" "Conjunto de dados do solo da bacia do reservatório do DNOS-CORSAN, Santa Maria, RS" "Compreende dados do solo de n = 400 observações feitas entre 2009 e 2013 na bacia do reservatório do Departamen"| __truncated__ "2.3" ...
#>   ..- attr(*, "spec")=List of 2
#>   .. ..$ cols   :List of 2
#>   .. .. ..$ campo: list()
#>   .. .. .. ..- attr(*, "class")= chr [1:2] "collector_character" "collector"
#>   .. .. ..$ valor: list()
#>   .. .. .. ..- attr(*, "class")= chr [1:2] "collector_character" "collector"
#>   .. ..$ default: list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_guess" "collector"
#>   .. ..- attr(*, "class")= chr "col_spec"
#>  $ data    :'data.frame':	400 obs. of  40 variables:
#>   ..$ dataset_id                 : chr [1:400] "ctb0003" "ctb0003" "ctb0003" "ctb0003" ...
#>   ..$ observacao_id              : chr [1:400] "sm-dnos-001" "sm-dnos-002" "sm-dnos-003" "sm-dnos-004" ...
#>   ..$ observacao_data            : chr [1:400] "10/02/2009" "10/02/2009" "10/02/2009" "10/02/2009" ...
#>   ..$ coord_sistema              : chr [1:400] "EPSG:4326" "EPSG:4326" "EPSG:4326" "EPSG:4326" ...
#>   ..$ coord_x                    : num [1:400] -53.8 -53.8 -53.8 -53.8 -53.8 ...
#>   ..$ coord_y                    : num [1:400] -29.7 -29.7 -29.7 -29.7 -29.7 ...
#>   ..$ coord_precisao             : num [1:400] 15 15 15 15 15 15 15 15 15 15 ...
#>   ..$ coord_fonte                : chr [1:400] "GPS" "GPS" "GPS" "GPS" ...
#>   ..$ pais_id                    : chr [1:400] "BR" "BR" "BR" "BR" ...
#>   ..$ estado_id                  : chr [1:400] "RS" "RS" "RS" "RS" ...
#>   ..$ municipio_id               : logi [1:400] NA NA NA NA NA NA ...
#>   ..$ amostra_tipo               : chr [1:400] "composta" "composta" "composta" "composta" ...
#>   ..$ amostra_quanti             : int [1:400] 3 3 3 3 3 3 3 3 3 3 ...
#>   ..$ amostra_area               : int [1:400] 100 100 100 100 100 100 100 100 100 100 ...
#>   ..$ taxon_sibcs_2009           : chr [1:400] "RY" "PBAC" "PBAC" "RY" ...
#>   ..$ terra_usoatual             : chr [1:400] "animal husbandry" "animal husbandry" "crop agriculture" "native forest" ...
#>   ..$ geo_parental               : chr [1:400] "sedimentary" "sedimentary" "sedimentary" "sedimentary" ...
#>   ..$ geo_formacao               : chr [1:400] NA NA NA NA ...
#>   ..$ geo_litologia              : chr [1:400] NA NA NA NA ...
#>   ..$ relevo_declividade         : int [1:400] 3 3 3 3 3 3 3 3 3 3 ...
#>   ..$ relevo_drenagem            : chr [1:400] "well" "moderately well" "moderately well" "well" ...
#>   ..$ geo_pedregosidade_k2       : chr [1:400] "Não pedregosa" "Não pedregosa" "Não pedregosa" "Não pedregosa" ...
#>   ..$ geo_rochosidade_k2         : chr [1:400] "Não rochosa" "Não rochosa" "Não rochosa" "Não rochosa" ...
#>   ..$ fito_cobertura             : chr [1:400] "medium" "high" "low" "low" ...
#>   ..$ observacao_autor           : chr [1:400] "Alessandro Samuel Rosa, Pablo Miguel, Andrisa Balbinot, Jean Michel Moura Bueno" "Alessandro Samuel Rosa, Pablo Miguel, Andrisa Balbinot, Jean Michel Moura Bueno" "Alessandro Samuel Rosa, Pablo Miguel, Andrisa Balbinot, Jean Michel Moura Bueno" "Alessandro Samuel Rosa, Pablo Miguel, Andrisa Balbinot, Jean Michel Moura Bueno" ...
#>   ..$ camada_id                  : chr [1:400] "1" "1" "1" "1" ...
#>   ..$ camada_nome                : chr [1:400] NA NA NA NA ...
#>   ..$ amostra_id                 : chr [1:400] NA NA NA NA ...
#>   ..$ profund_sup                : chr [1:400] "0" "0" "0" "0" ...
#>   ..$ profund_inf                : chr [1:400] "20" "20" "20" "20" ...
#>   ..$ areia_naoh_esferas_peneira : int [1:400] 912 735 753 815 684 549 488 792 857 788 ...
#>   ..$ argila_naoh_esferas_pipeta : int [1:400] 44 102 99 82 100 156 169 102 68 91 ...
#>   ..$ calcio_kcl_eaa             : num [1:400] 3 3.2 4.3 4.4 3.8 3.6 4 2.5 1.3 2.7 ...
#>   ..$ magnesio_kcl_eaa           : num [1:400] 1 1 1.6 1.7 0.8 0.9 1 0.5 0.3 0.5 ...
#>   ..$ aluminio_kcl_naoh          : num [1:400] 0.1 0.8 0.2 0.3 0.6 1.3 1.1 1.4 1.4 0.3 ...
#>   ..$ potassio_mehlich1_eeac     : num [1:400] 0.2 0.2 0.6 0.5 0.2 0.2 0.2 0.1 0.1 0.1 ...
#>   ..$ sodio_mehlich1_eeac        : num [1:400] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ carbono_cromo_30min150_mohr: int [1:400] 5 8 10 19 13 11 13 12 9 11 ...
#>   ..$ dsi_cilindro               : num [1:400] 1.49 1.56 1.3 1.15 1.17 1.38 1.41 1.07 1.23 1.57 ...
#>   ..$ cascalho_olho              : int [1:400] NA NA NA NA NA NA NA NA NA NA ...
#>  $ metadata:'data.frame':	40 obs. of  14 variables:
#>   ..$ dataset_id       : Factor w/ 1 level "ctb0003": 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ tabela_id        : chr [1:40] "observacao" "observacao" "observacao" "observacao" ...
#>   ..$ campo_id         : chr [1:40] "observacao_id" "observacao_data" "coord_sistema" "coord_x" ...
#>   ..$ campo_nome       : chr [1:40] "Identificação da observação" "Data da observação" "Sistema de referência de coordenadas" "Longitude" ...
#>   ..$ campo_descricao  : chr [1:40] "Identificador exclusivo da observação do solo." "Data da observação do solo no formato dia/mês/ano." "Sistema de referência de coordenadas espaciais utilizado (WGS 84)." "Coordenadas geográfica (latitude) em graus." ...
#>   ..$ campo_fator      : num [1:40] NA NA NA 1 1 1 NA NA NA NA ...
#>   ..$ campo_unidade    : chr [1:40] NA NA NA "°" ...
#>   ..$ campo_tipo       : chr [1:40] "Texto" "Data" "Inteiro" "Real" ...
#>   ..$ lab_nome         : chr [1:40] "n/a" "n/a" "n/a" "n/a" ...
#>   ..$ lab_pais_id      : chr [1:40] "n/a" "n/a" "n/a" "n/a" ...
#>   ..$ lab_municipio_id : chr [1:40] "n/a" "n/a" "n/a" "n/a" ...
#>   ..$ lab_codigo_postal: chr [1:40] "n/a" "n/a" "n/a" "n/a" ...
#>   ..$ lab_rua_nome     : chr [1:40] "n/a" "n/a" "n/a" "n/a" ...
#>   ..$ lab_rua_numero   : chr [1:40] "n/a" "n/a" "n/a" "n/a" ...