Write data to an Excel workbook

Write data downloaded from the Free Brazilian Repository for Open Soil Data -- febr, http://www.ufsm.br/febr -- to an Excel workbook.

febr2xlsx(x, file, row.names = FALSE, ...)

Arguments

x
A data.frame or list of data.frames to write to the Excel workbook.

file
Character string indicating the path to the output XLSX file.

row.names
(optional) Logical value indicating whether the row names of x are to be written along with x to the file.

...
(optional) Further arguments passed to write.xlsx.

Examples

dts <-
  febr::febr(dataset = "ctb0013",
             variable = "all",
             merge = TRUE,
             progress = FALSE, verbose = FALSE)
febr2xlsx(x = dts, file = tempfile(fileext = ".xlsx"))