Go to FEBR

Visit FEBR Soil Data Repository assets on the web.

goto(data.set, page)

Arguments

data.set

(optional) Character vector indicating a (unique) dataset whose metadata web page you wish to visit.

page

(optional) Character string indicating a web page of the FEBR Soil Data Repository, with options:

"febr": FEBR main web page,

"dictionary": FEBR data dictionary and vocabulary,

"forum": FEBR public forum at Google Groups,

"github": febr package source code repository on GitHub,

"index": FEBR data set index,

"search": data set search web page,

"package": febr package on CRAN,

"template": FEBR spreadsheet template on Google Sheets,

"units": units and conversion factors used in FEBR.

Value

Load a given URL into an HTML browser via utils::browseURL().

See also

Author

Alessandro Samuel-Rosa alessandrosamuelrosa@gmail.com

Examples

if (interactive()) {
# Go to the FEBR web page
goto(page = "febr")

# Go to the data set index
goto(page = "index")

# Go to the GitHub repository
goto(page = "github")
}