Get 'identification' table

Download data from the 'identification' ("identificacao") table of one or more soil datasets published in the Free Brazilian Repository for Open Soil Data (FEBR), https://www.pedometria.org/febr/. This table includes data such as dataset title and description, author and institution, data license, and much more.

identification(data.set, progress = TRUE, verbose = TRUE, febr.repo = NULL)

Arguments

data.set
Character vector indicating the identification code of one or more data sets. Use data.set = "all" to download all data sets.

progress
(optional) Logical value indicating if a download progress bar should be displayed.

verbose
(optional) Logical value indicating if informative messages should be displayed. Generally useful to identify datasets with inconsistent data. Please report to febr-forum@googlegroups.com if you find any issue.

febr.repo
(optional) Defaults to the remote file directory of the Federal University of Technology - Paraná at https://cloud.utfpr.edu.br/index.php/s/Df6dhfzYJ1DDeso. Alternatively, a local directory path can be informed if the user has a local copy of the data repository.

Value

A list of data frames or a data frame with data of the chosen dataset(s).

Note

Check the new core data download function readFEBR().

Author

Alessandro Samuel-Rosa alessandrosamuelrosa@gmail.com

Examples

# \donttest{
# res <- identification(data.set = c("ctb0003", "ctb0000"))
res <- metadata(data.set = c("ctb0003", "ctb0002"))

#> 
  |                                                                            
  |                                                                      |   0%

#> 
#> Reading ctb0003-metadado...

#> 
  |                                                                            
  |===================================                                   |  50%

#> 
#> Reading ctb0002-metadado...

#> 
  |                                                                            
  |======================================================================| 100%

# }