Get soil data

Download soil data from one or more data sets published in the Data Repository of the Brazilian Soil.

readFEBR(
  data.set,
  data.table = c("identificacao", "versionamento", "metadado", "observacao", "camada"),
  febr.repo = NULL,
  verbose = TRUE,
  ...
)

Arguments

data.set

Character vector indicating the identification code of one or more data sets. Use data.set = "all" to download all data sets.

data.table

Character vector indicating one or more data tables, with supported values "identificacao", "versionamento", "metadado", "observacao", and "camada".

febr.repo

(optional) Defaults to the remote file directory of the Federal University of Technology - Paraná at https://cloud.utfpr.edu.br/index.php/s/Df6dhfzYJ1DDeso. Alternatively, a local directory path can be informed if the user has a local copy of the data repository.

verbose

(optional) Logical value indicating if informative messages should be displayed. Generally useful to identify issues—please report to febr-forum@googlegroups.com if you find any.

...

(optional) Arguments passed to data.table::fread().

Value

A list of data tables (data frames) with data from the chosen data sets.

Autor

Alessandro Samuel-Rosa alessandrosamuelrosa@gmail.com

Examples

res <- readFEBR(data.set = "ctb0003")
#> Reading...
#> https://cloud.utfpr.edu.br/index.php/s/Df6dhfzYJ1DDeso/download?path=%2Fctb0003&files=ctb0003-identificacao.txt
#> https://cloud.utfpr.edu.br/index.php/s/Df6dhfzYJ1DDeso/download?path=%2Fctb0003&files=ctb0003-versionamento.txt
#> https://cloud.utfpr.edu.br/index.php/s/Df6dhfzYJ1DDeso/download?path=%2Fctb0003&files=ctb0003-metadado.txt
#> https://cloud.utfpr.edu.br/index.php/s/Df6dhfzYJ1DDeso/download?path=%2Fctb0003&files=ctb0003-observacao.txt
#> https://cloud.utfpr.edu.br/index.php/s/Df6dhfzYJ1DDeso/download?path=%2Fctb0003&files=ctb0003-camada.txt