Process soil taxonomy

Extract and process soil taxonomic data from textual soil classification descriptions of the Brazilian Soil Classification System (SiBCS).

taxonomy(
  text,
  method = "decompose",
  sep = " ",
  pattern = c(", ", " A ", " textura ")
)

Arguments

text

Character string(s) with soil classification description(s) (in Portuguese).

method

Character string defining the string processing method. Options:

"decompose": decompose the Brazilian soil classification into its four higher categorical levels (order, suborder, large group, and subgroup).

sep

Character string. Defaults to sep = " ".

pattern

Character string (in Portuguese). Defaults to pattern = c(", ", " A ", " textura ").

Value

An object of class base::data.frame with four named columns: ordem (UPPER CASE), subordem (UPPER CASE), grangrupo (Title Case), and subgrupo (lower case).

Referências

Santos, H. G. dos, Jacomine, P. K. T., Anjos, L. H. C. dos, Oliveira, V. Á. de, Lumbreras, J. F., Coelho, M. R., Almeida, J. A. de, Araújo Filho, J. C. de, Oliveira, J. B. de, & Cunha, T. J. F. (2018). Sistema Brasileiro de Classificação de Solos (5th ed., p. 531). Embrapa. https://www.infoteca.cnptia.embrapa.br/infoteca/handle/doc/1094003.

IBGE. (2015). Manual Técnico de Pedologia (3rd ed., p. 430). Instituto Brasileiro de Geografia e Estatística. https://biblioteca.ibge.gov.br/visualizacao/livros/liv95017.pdf

Autor

Alessandro Samuel-Rosa alessandrosamuelrosa@gmail.com

Examples

text <-
  c("CAMBISSOLO HÁPLICO Ta Eutrófico léptico A proeminente textura média",
    "PLANOSSOLO HÁPLICO Distrófico solódico êndico plintossólico, textura média, Tb",
    "CHERNOSSOLO ARGILÚVICO Órtico típico textura média com cascalho/argilosa com cascalho",
    "ARGISSOLO VERMELHO-AMARELO",
    "Latossolo"
  )
taxonomy(text)
#>         ordem         subordem    grangrupo                      subgrupo
#> 1  CAMBISSOLO          HÁPLICO Ta Eutrófico                       léptico
#> 2  PLANOSSOLO          HÁPLICO   Distrófico solódico êndico plintossólico
#> 3 CHERNOSSOLO       ARGILÚVICO       Órtico                        típico
#> 4   ARGISSOLO VERMELHO-AMARELO         <NA>                          <NA>
#> 5   LATOSSOLO             <NA>         <NA>                          <NA>