Get units table

Download data from the units ("unidades") table of the Free Brazilian Repository for Open Soil Data -- FEBR, https://www.pedometria.org/febr/. This includes measurement units and conversion factors. This is used to standardize the data contained in a dataset when downloading it via layer or observation.

unit(source, target)

Arguments

source
(optional) Character vector indicating one or more source measurement units.

target
(optional) Character vector indicating one or more target measurement units.

Value

A data.frame with source and target measurement units and their corresponding conversion factors.

References

Teixeira, P. C., Donagemma, G. K., Fontana, A., Teixeira, W. G. (2017) Manual de Métodos de Análise de Solo. Brasília: Embrapa.

See also

Author

Alessandro Samuel-Rosa alessandrosamuelrosa@gmail.com

Examples

#res <- unit(source = c("%", "dag/kg"), target = "g/kg")
#res