Access the FEBR dictionary

Download definitions from the dictionary of the Data Repository of the Brazilian Soil. For each field, the dictionary includes an identification code and name, the standard measurement unit, recommended number of decimal places, type of data, and description of the respective analytical method. The dictionary is used to standardize the data contained in a dataset.

dictionary(table, variable, unit, precision, active = NULL)

Arguments

table

(optional) Character vector indicating one or more table IDs that should be used to filter the dictionary. Accepted values: "metadado", "versionamento", "observacao", and "camada".

variable

(optional) Character vector indicating one or more variables that should be used to filter the dictionary. Accepts both specific identification codes, e.g. "ferro_oxalato_icpoes", as well as general identification codes, e.g. "ferro".

unit

(optional) Character vector indicating one or more measurement units that should be used to filter the dictionary. For example, ' "g/kg", "g/cm^3", and "cmolc/kg".

precision

(optional) Integer vector indicating one or more number of decimal places that should be used to filter the dictionary.

active

(optional) Logical value indicating whether active (active = TRUE), inactive (active = FALSE) or any (active = NULL, default) field should be returned.

Value

An object of class data.frame with definitions for the selected fields.

Referências

Teixeira, P. C., Donagemma, G. K., Fontana, A., Teixeira, W. G. (2017) Manual de Métodos de Análise de Solo. Brasília: Embrapa.

See also

Autor

Alessandro Samuel-Rosa alessandrosamuelrosa@gmail.com

Examples

res <- dictionary(variable = "ferro")
head(res)
#>   tabela_id               campo_id campo_vital campo_oldid
#> 1    camada  ferro_aquaregia_icpms       FALSE        <NA>
#> 2    camada ferro_aquaregia_icpoes       FALSE        <NA>
#> 3    camada   ferro_cloridrico_eaa       FALSE        <NA>
#> 4    camada    ferro_ditionito_eaa       FALSE        <NA>
#> 5    camada    ferro_ditionito_eam       FALSE        <NA>
#> 6    camada ferro_ditionito_icpoes       FALSE        <NA>
#>                campo_nome
#> 1             Ferro total
#> 2             Ferro total
#> 3             Ferro total
#> 4 Ferro livre (extraível)
#> 5 Ferro livre (extraível)
#> 6 Ferro livre (extraível)
#>                                                                                                                                                                                                                                                                                                                                                                                                                              campo_descricao
#> 1                                                                         Conteúdo de ferro total na fração terra fina seca ao ar (< 2 mm). Digestão à quente com ácido nítrico (HNO3) + ácido clorídrico (HCl) concentrados, numa proporção de 3:1 [aquaregia]. {ESPECIFICAR FORMA DE AQUECIMENTO, TEMPO E TEMPERATURA} Quantificação por espectrometria de massa com plasma acoplado indutivamente [icpms]. {INDICAR REFERÊNCIA DO MÉTODO}
#> 2                                                              Conteúdo de ferro total na fração terra fina seca ao ar (< 2 mm). Digestão à quente com ácido nítrico (HNO3) + ácido clorídrico (HCl) concentrados, numa proporção de 3:1 [aquaregia]. {ESPECIFICAR FORMA DE AQUECIMENTO, TEMPO E TEMPERATURA} Quantificação por espectrometria de emissão atômica com plasma acoplado indutivamente [icpoes]. {INDICAR REFERÊNCIA DO MÉTODO}
#> 3                                                                                                                                                           Conteúdo de ferro total na fração terra fina seca ao ar (< 2 mm). Digestão à quente com ácido clorídrico (HCl) 6 mol/L [cloridrico]. {ESPECIFICAR TEMPERATURA E TEMPO DE AQUECIMENTO} Quantificação por espectrometria de absorção atômica [eaa]. {INDICAR REFERÊNCIA DO MÉTODO}
#> 4                                     Conteúdo de ferro livre (extraível) na fração terra fina seca ao ar (< 2 mm). Extração com citrato de sódio (C6H5O7NA3.2H2O) 0,3 mol/L + bicarbonato de sódio (NaHCO3) 1 mol/L + ditionito de sódio [ditionito] sob aquecimento em banho-maria à temperatura de 75-80ºC. {ESPECIFICAR TEMPO DE AQUECIMENTO} Quantificação por espectrometria de absorção atômica [eaa]. {INDICAR REFERÊNCIA DO MÉTODO}
#> 5                Conteúdo de ferro livre (extraível) na fração terra fina seca ao ar (< 2 mm). Extração com citrato de sódio (C6H5O7NA3.2H2O) 0,3 mol/L + bicarbonato de sódio (NaHCO3) 1 mol/L + ditionito de sódio [ditionito] sob aquecimento em banho-maria à temperatura de 75-80ºC. {ESPECIFICAR TEMPO DE AQUECIMENTO} Quantificação por espectrofotometria de absorção molecular (colorimetria) [eam]. {INDICAR REFERÊNCIA DO MÉTODO}
#> 6 Conteúdo de ferro livre (extraível) na fração terra fina seca ao ar (< 2 mm). Extração com citrato de sódio (C6H5O7NA3.2H2O) 0,3 mol/L + bicarbonato de sódio (NaHCO3) 1 mol/L + ditionito de sódio [ditionito] sob aquecimento em banho-maria à temperatura de 75-80ºC. {ESPECIFICAR TEMPO DE AQUECIMENTO} Quantificação por espectrometria de emissão atômica com plasma acoplado indutivamente [icpoes]. {INDICAR REFERÊNCIA DO MÉTODO}
#>   campo_opcao campo_unidade campo_precisao campo_tipo campo_cat campo_carater
#> 1           -          g/kg              1       real   química      opcional
#> 2           -          g/kg              1       real   química      opcional
#> 3           -          g/kg              1       real   química      opcional
#> 4           -          g/kg              2       real   química      opcional
#> 5           -          g/kg              2       real   química      opcional
#> 6           -          g/kg              2       real   química      opcional
#>   campo_ativo basica pedologica campo_exemplo
#> 1           1      -          -          <NA>
#> 2           1      -          -          <NA>
#> 3           1      -          -          <NA>
#> 4           1      -          -          <NA>
#> 5           1      -          -          <NA>
#> 6           1      -          -          <NA>