Go to FEBR

Go to one of the web pages of the Free Brazilian Repository for Open Soil Data (FEBR), including project and dataset web pages.

goto(data.set, page)

Arguments

data.set
(optional) Character vector indicating a single data set.

page
(optional) Character string indicating a web page of the FEBR. Options are: "febr" (FEBR web page), "search" (data set search web page), "package" (febr package on CRAN), "github" (febr package source code repository on GitHub), "gitlab" (FEBR source code repository on GitLab), "forum" (FEBR public forum at Google Groups), "units" (units and conversion factors used in FEBR), "dictionary" (dictionary used in FEBR), "index" (data set index), and "template" (spreadsheet data template).

Author

Alessandro Samuel-Rosa alessandrosamuelrosa@gmail.com

Examples

if (interactive()) {
# Go to the FEBR web page
goto(page = "febr")

# Go to the data set index
goto(page = "index")

# Go to the GitHub repository
goto(page = "github")
}